create datasource common type
